+++ iblqt/__init__.py
@@ -1,3 +1,3 @@
 """A collection of extensions to the Qt framework."""
 
-__version__ = '0.1.2'
+__version__ = '0.2.0'