+++ iblqt/widgets.py
@@ -1,7 +1,111 @@
 """Graphical user interface components."""
 
-from qtpy.QtWidgets import QPushButton
-from qtpy.QtCore import Signal, Slot, Property
+from typing import Any
+
+from qtpy.QtWidgets import (
+    QPushButton,
+    QStyledItemDelegate,
+    QStyleOptionButton,
+    QCheckBox,
+    QStyle,
+    QApplication,
+    QStyleOptionViewItem,
+)
+from qtpy.QtCore import (
+    Signal,
+    Slot,
+    Property,
+    QRect,
+    QEvent,
+    QModelIndex,
+    QAbstractItemModel,
+)
+from qtpy.QtGui import QPainter, QMouseEvent
+
+
+class CheckBoxDelegate(QStyledItemDelegate):
+    """
+    A custom delegate for rendering checkboxes in a QTableView or similar widget.
+
+    This delegate allows for the display and interaction with boolean data as checkboxes.
+    """
+
+    def paint(
+        self, painter: QPainter, option: QStyleOptionViewItem, index: QModelIndex
+    ) -> None:
+        """
+        Paints the checkbox in the view.
+
+        Parameters
+        ----------
+        painter : QPainter
+            The painter used to draw the checkbox.
+        option : QStyleOptionButton
+            The style option containing the information needed for painting.
+        index : QModelIndex
+            The index of the item in the model.
+        """
+        super().paint(painter, option, index)
+        control = QStyleOptionButton()
+        control.rect = QRect(option.rect.topLeft(), QCheckBox().sizeHint())
+        control.rect.moveCenter(option.rect.center())
+        control.state = QStyle.State_On if index.data() is True else QStyle.State_Off
+        QApplication.style().drawControl(
+            QStyle.ControlElement.CE_CheckBox, control, painter
+        )
+
+    def displayText(self, value: Any, locale: Any) -> str:
+        """
+        Return an empty string to hide the text representation of the data.
+
+        Parameters
+        ----------
+        value : Any
+            The value to be displayed (not used).
+        locale : Any
+            The locale to be used (not used).
+
+        Returns
+        -------
+        str
+            An empty string.
+        """
+        return ''
+
+    def editorEvent(
+        self,
+        event: QEvent,
+        model: QAbstractItemModel,
+        option: QStyleOptionViewItem,
+        index: QModelIndex,
+    ) -> bool:
+        """
+        Handle user interaction with the checkbox.
+
+        Parameters
+        ----------
+        event : QEvent
+            The event that occurred (e.g., mouse click).
+        model : QAbstractItemModel
+            The model associated with the view.
+        option : QStyleOptionViewItem
+            The style option containing the information needed for handling the event.
+        index : QModelIndex
+            The index of the item in the model.
+
+        Returns
+        -------
+        bool
+            True if the event was handled, False otherwise.
+        """
+        if isinstance(event, QMouseEvent) and event.type() == QEvent.MouseButtonRelease:
+            checkbox_rect = QRect(option.rect.topLeft(), QCheckBox().sizeHint())
+            checkbox_rect.moveCenter(option.rect.center())
+            if checkbox_rect.contains(event.pos()):
+                model.setData(index, not model.data(index))
+                event.accept()
+                return True
+        return super().editorEvent(event, model, option, index)
 
 
 class StatefulButton(QPushButton):