+++ CHANGELOG.md
@@ -7,6 +7,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] - 2024-10-08
+
+### Added
+
+- widgets.CheckBoxDelegate: render checkboxes in a QTableView or similar widget
+- start adding unit tests
+
+### Changed
+
+- core.DataFrameTableModel: stop use of QVariant
+- core.DataFrameTableModel: set dataFrame _after_ connecting signals in initialization
+- core.DataFrameTableModel: default to horizontal orientation
+
 ## [0.1.2] - 2024-10-01
 
 ### Changed
@@ -36,6 +49,7 @@ _First release._
 - core.ColoredDataFrameTableModel: An extension of DataFrameTableModel providing color-mapped numerical data.
 - widgets.StatefulButton: A QPushButton that maintains an active/inactive state.
 
+[0.2.0]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.2.0
 [0.1.2]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.1.2
 [0.1.1]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.1.1
 [0.1.0]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.1.0